implemented msb for compressor

+++ src/compressor/dictionary/trie-on-sparse-array/main.h
@@ -24,7 +24,7 @@ LZWS_INLINE void lzws_compressor_initialize_dictionary(lzws_compressor_dictionar
 }
 
 LZWS_INLINE lzws_result_t lzws_compressor_allocate_dictionary(lzws_compressor_dictionary_t* dictionary, uint8_t max_code_bits) {
-  uint32_t total_codes = LZWS_POWERS_OF_TWO[max_code_bits];
+  size_t total_codes = LZWS_POWERS_OF_TWO[max_code_bits];
 
   lzws_code_t* codes = lzws_allocate_array(sizeof(lzws_code_t), total_codes * LZWS_ALPHABET_LENGTH, true, LZWS_UNDEFINED_NEXT_CODE);
   if (codes == NULL) {
@@ -37,18 +37,18 @@ LZWS_INLINE lzws_result_t lzws_compressor_allocate_dictionary(lzws_compressor_di
 }
 
 LZWS_INLINE void lzws_compressor_clear_dictionary(lzws_compressor_dictionary_t* dictionary, uint8_t max_code_bits) {
-  uint32_t total_codes = LZWS_POWERS_OF_TWO[max_code_bits];
+  size_t total_codes = LZWS_POWERS_OF_TWO[max_code_bits];
 
   lzws_fill_array(dictionary->codes, sizeof(lzws_code_t), total_codes * LZWS_ALPHABET_LENGTH, LZWS_UNDEFINED_NEXT_CODE);
 }
 
 LZWS_INLINE lzws_code_t lzws_compressor_get_next_code_from_dictionary(lzws_compressor_dictionary_t* dictionary, lzws_code_t current_code, uint8_t symbol) {
-  uint32_t code_index = current_code * LZWS_ALPHABET_LENGTH + symbol;
+  size_t code_index = current_code * LZWS_ALPHABET_LENGTH + symbol;
   return dictionary->codes[code_index];
 }
 
 LZWS_INLINE void lzws_compressor_save_next_code_to_dictionary(lzws_compressor_dictionary_t* dictionary, lzws_code_t current_code, uint8_t symbol, lzws_code_t code) {
-  uint32_t code_index           = current_code * LZWS_ALPHABET_LENGTH + symbol;
+  size_t code_index             = current_code * LZWS_ALPHABET_LENGTH + symbol;
   dictionary->codes[code_index] = code;
 }
